[Engine] Start processing of a specified project now returns promise.
This allow to send different initial requests to the server.
New field `Query` has been also added to msg.CommandLine request.
Server uses it to determine a sequence of passes to run.
For example, transformations would be run if Query is some TSAR transformation options.
If Query is not specified than default server pass sequence is used to show analysis results.

Signed-off-by: Nikita Kataev <kaniandr@gmail.com>